Mostly done with review
